Refactor ThreeEarthSystem to modular architecture and clean up HTML/JS scope
- Split `three-earth-system.js` into `earth/` modules: `config`, `scene`, `assets`, `camera`, `stars`, `ui`.
- Clean up `index.html` by removing duplicate meta tags, scripts, and preloads.
- Refactor `main.js` to minimize global scope pollution and improve test detection via `ENV` object.
- Update `SectionTracker` in `shared-utilities.js` to handle dynamic footer loading gracefully and suppress initial warnings.
- Fix memory leak in `PerformanceMonitor` cleanup.
- Ensure safe timer cleanup in `CameraManager`.
- Apply Prettier formatting to all changed files.
