+++ content/webentwicklung/particles/earth/config.js
@@ -1,0 +1,151 @@
+export const CONFIG = {
+  EARTH: {
+    RADIUS: 3.5,
+    SEGMENTS: 64,
+    BUMP_SCALE: 0.008,
+    EMISSIVE_INTENSITY: 0.2,
+    EMISSIVE_PULSE_SPEED: 0.3,
+    EMISSIVE_PULSE_AMPLITUDE: 0.08
+  },
+  CLOUDS: {
+    ALTITUDE: 0.03,
+    ROTATION_SPEED: 0.0008,
+    OPACITY: 0.3
+  },
+  ATMOSPHERE: {
+    SCALE: 1.015,
+    FRESNEL_POWER: 4.5,
+    RAYLEIGH_SCALE: 1.028,
+    MIE_SCALE: 1.018,
+    RAYLEIGH_COLOR: 0x4488ff,
+    MIE_COLOR: 0xffbb66,
+    RAYLEIGH_INTENSITY: 0.08,
+    MIE_INTENSITY: 0.04,
+    SCATTERING_STRENGTH: 0.18
+  },
+  OCEAN: {
+    SHININESS: 100.0,
+    SPECULAR_INTENSITY: 0.5,
+    SPECULAR_COLOR: 0xffffff
+  },
+  SUN: {
+    RADIUS: 15,
+    HEIGHT: 3.0,
+    INTENSITY: 1.8
+  },
+  LIGHTING: {
+    DAY: {
+      AMBIENT_INTENSITY: 1.4,
+      AMBIENT_COLOR: 0x606060,
+      SUN_INTENSITY: 1.8
+    },
+    NIGHT: {
+      AMBIENT_INTENSITY: 0.3,
+      AMBIENT_COLOR: 0x202845,
+      SUN_INTENSITY: 0.35
+    }
+  },
+  STARS: {
+    COUNT: 3000,
+    TWINKLE_SPEED: 0.3,
+    ANIMATION: {
+      DURATION: 3500,
+      CAMERA_SETTLE_DELAY: 2200,
+      MIN_UPDATE_INTERVAL: 100,
+      CARD_FADE_START: 0.7,
+      CARD_FADE_END: 0.95,
+      SPREAD_XY: 0.6,
+      SPREAD_Z: 0.3,
+      LERP_FACTOR: 0.08
+    }
+  },
+  MOON: {
+    RADIUS: 0.95,
+    DISTANCE: 25,
+    ORBIT_SPEED: 0.00025,
+    SEGMENTS: 48,
+    BUMP_SCALE: 0.015
+  },
+  CAMERA: {
+    FOV: 45,
+    NEAR: 0.1,
+    FAR: 1000,
+    ZOOM_MIN: 5,
+    ZOOM_MAX: 25,
+    LERP_FACTOR: 0.06,
+    PRESETS: {
+      hero: {
+        x: -6.5,
+        y: 4.8,
+        z: 10.5,
+        lookAt: { x: 0, y: -0.5, z: 0 }
+      },
+      features: {
+        x: 7.0,
+        y: 5.5,
+        z: 7.5,
+        lookAt: { x: 0, y: 0.5, z: 0 }
+      },
+      about: {
+        x: -3.2,
+        y: 3.0,
+        z: 9.5,
+        lookAt: { x: 0, y: 0, z: 0 }
+      },
+      contact: {
+        x: -0.5,
+        y: 3.0,
+        z: 9.5,
+        lookAt: { x: 0, y: 0, z: 0 }
+      }
+    },
+    TRANSITION_DURATION: 1.8
+  },
+  SHOOTING_STARS: {
+    BASE_FREQUENCY: 0.003,
+    SHOWER_FREQUENCY: 0.02,
+    SHOWER_DURATION: 180,
+    SHOWER_COOLDOWN: 1200,
+    MAX_SIMULTANEOUS: 3
+  },
+  PERFORMANCE: {
+    PIXEL_RATIO: Math.min(window.devicePixelRatio, 2.0),
+    TARGET_FPS: 55,
+    DRS_DOWN_THRESHOLD: 48,
+    DRS_UP_THRESHOLD: 58
+  },
+  QUALITY_LEVELS: {
+    HIGH: {
+      minFPS: 48,
+      multiLayerAtmosphere: true,
+      oceanReflections: true,
+      cloudLayer: true,
+      meteorShowers: true
+    },
+    MEDIUM: {
+      minFPS: 28,
+      multiLayerAtmosphere: false,
+      oceanReflections: true,
+      cloudLayer: true,
+      meteorShowers: true
+    },
+    LOW: {
+      minFPS: 0,
+      multiLayerAtmosphere: false,
+      oceanReflections: false,
+      cloudLayer: false,
+      meteorShowers: false
+    }
+  },
+  PATHS: {
+    TEXTURES: {
+      DAY: '/content/img/earth/textures/earth_day.webp',
+      NIGHT: '/content/img/earth/textures/earth_night.webp',
+      NORMAL: '/content/img/earth/textures/earth_normal.webp',
+      BUMP: '/content/img/earth/textures/earth_bump.webp',
+      CLOUDS: '/content/img/earth/textures/earth_clouds_1024.png',
+      MOON: '/content/img/earth/textures/moon_texture.webp',
+      MOON_BUMP: '/content/img/earth/textures/moon_bump.webp'
+    }
+  }
+};
